add GalleryPreviewSection component
