+++ src/components/organisms/GalleryPreviewSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './GalleryPreviewSection'
